+++ src/main/java/racingcar/dao/DummyGameResultDao.java
@@ -6,12 +6,12 @@ import java.util.List;
 
 public class DummyGameResultDao implements GameResultDao {
     @Override
-    public int save(GameResult gameResult) {
+    public long save(GameResult gameResult) {
         return 0;
     }
 
     @Override
-    public GameResult findById(int id) {
+    public GameResult findById(long id) {
         return null;
     }
